add field under Header, move Search function to new field
