+++ pages/index.js
@@ -1,45 +1,16 @@
 import Project from "@/components/Project";
 import styled from "styled-components";
-import { useState } from "react";
-import Fuse from "fuse.js";
+import SearchField from "@/components/SearchField";
 
-export default function HomePage({ projects, onToggleFavorite, isHidden }) {
-  const [query, setQuery] = useState("");
-
-  const fuse = new Fuse(projects, {
-    keys: ["title", "description", "materials", "steps.desc"],
-    includeScore: true,
-    threshold: 0.3,
-    shouldSort: true,
-    ignoreLocation: true,
-    ignoreFieldNorm: true,
-  });
-  const results = fuse.search(query);
-  const searchResults = query ? results.map((result) => result.item) : projects;
-
-  function handleSearch(event) {
-    const { value } = event.target;
-    setQuery(value);
-  }
-
+export default function HomePage({
+  handleSearch,
+  query,
+  searchResults,
+  onToggleFavorite,
+}) {
   return (
     <>
-      <StyledSearchField>
-        <label htmlFor="search" hidden={isHidden}>
-          Search
-        </label>
-        <input
-          type="search"
-          hidden={isHidden}
-          id="search"
-          name="search"
-          size={30}
-          placeholder="... for title, material etc."
-          value={query}
-          onChange={handleSearch}
-        />
-      </StyledSearchField>
-
+      <SearchField handleSearch={handleSearch} query={query} />
       <StyledSection>
         {searchResults.map((project) => (
           <Project
@@ -66,11 +37,4 @@ const StyledSection = styled.section`
   margin: 10px;
   background-color: var(--background-color);
   border-radius: 20px;
-`;
-
-const StyledSearchField = styled.article`
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
 `;
